test(core): keep the envelope helpers in the helper file
`mustOmitBundledScheme` and `keysOf` were defined inside the test-case files
that first needed them, against this module's convention of one case per file
with shared helpers in `helpers_*_test.go`; the second case to need either
would have had to import from a case file or duplicate it.

Also corrects a doc comment that still named `sourceFileKey`, the local keying
helper this cycle replaced with `driver.TransformOutputKey`.
